# openapi.yaml
openapi: 3.0.3
info:
  title: CivicOS API
  version: 1.0.0
servers:
  - url: https://civicos.onrender.com
paths:
...
  /api/friends:
    get:
      summary: Get friends list
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /api/friends/requests:
    get:
      summary: Get pending friend requests
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/friends/request:
    post:
      summary: Send friend request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                toUserId: { type: string }
              required: [toUserId]
      responses:
        '200': { description: OK }
  /api/friends/accept:
    post:
      summary: Accept friend request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requestId: { type: integer }
              required: [requestId]
      responses:
        '200': { description: OK }
...
  /api/admin/moderation-dashboard:
    get:
      summary: Combined moderation dashboard (summary + recent posts/comments)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        '200': { description: OK }
  /api/admin/summary:
    get:
      summary: Platform summary for admin dashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT